pkcs8: add Ed25519 examples; fix AlgorithmIdentifier parsing
Adds OpenSSL-generated examples of Ed25519 private and public keys in
both DER and PEM format.

Also adds support for the `parameters` field of `AlgorithmIdentifier`
being absent, as it is in Ed25519 keys. Previously only an explicit
ASN.1 NULL was supported (as used by RSA).

The encoding of Ed25519 keys does not yet match the OpenSSL-generated
keys, however I will address that in a focused follow-up PR.