pkcs8: special case RSA AlgorithmIdentifier parameters encoding
The immediate goal of this change is to ensure Ed25519 keys are encoded
with AlgorithmIdentifiers that match keys generated by OpenSSL.

The `parameters` field of `AlgorithmIdentifier` is defined as being
optional, and Ed25519 keys take advantage of this and completely omit
any sort of `parameters` from the ASN.1 SEQUENCE.

The previous encoding logic hardcoded the RSA encoding behavior defined
in RFC 3279 Section 2.3.1:

> The parameters field MUST have ASN.1 type NULL for this
> algorithm identifier.

However, this encoding appears to be a sort of unique historical
accident specific to RSA.

While we could pursue a solution that doesn't require special casing
RSA, it would be a breaking API change at the very least. It would also
complicate the API for (all?) non-RSA algorithms.