pkcs8: add `AlgorithmParameters` enum
This commit replaces the previous special casing of RSA `AlgorithmInfo`
with a more general and extensible solution.

The `parameters` field of `AlgorithmInfo` is now modeled as an optional
enum, which presently has `Null` and `Oid` variants.

With this, the RFC 3279 requirement that RSA's `AlgorithmIdentifier` has
an explicit NULL value can be directly modeled via the public API
without special casing.

It also leaves the door open to support other ASN.1 objects besides OIDs
as `parameters`, if that case ever arises.