const-oid: add `Nodes` iterator
Replaces the previous `AsRef<[u32]>` impl on `ObjectIdentifier` with an
iterator type: `Nodes`.

This allows for using a more compact internal representation for the
`ObjectIdentifier` struct, because it no longer requires that the first
two nodes are represented as u32 (they can be encoded in a single byte).

This is preliminary work for moving to such a layout.