const-oid: layout optimization
Stores the first two arcs of an OID as a `u8`, rather than `[u32; 2]`.

The range of values of these nodes is constrained such that this is
possible. The ASN.1 DER encoding takes advantage of this optimization.

With this change, it's possible to store an additional two arcs in the
same amount of space on 64-bit architectures (48-bytes).