der: initial crate (#153)
The `pkcs8` crate contained a rudimentary ASN.1 DER parser which was
successfully able to decode and encode a set of example files generated
by OpenSSL.

This code is both generally and immediately useful in a separate crate,
namely it would be nice to be able to use it to implement decoding and
encoding of ECC PKCS#8 keys (contained within `PrivateKeyInfo`), which
are an inner ASN.1 ANY type.

This commit contains an initial extraction of the DER-specific code from
the `pkcs8` crate into a new `der` crate.

This is not intended to be anything close to a final API, and there are
presently no tests other than indirectly exercising the code via the
`pkcs8` crate's tests. However, it compiles, and those tests pass, so
it's a first step towards a new crate.

There are many ways in which this code could be significantly improved
which I intend to work on as follow-up commits.