der: refactor decoder using lambdas
Adds a lambda argument to the `nested` function, having it perform TLV
decoding and yielding the decoded tag and value to the lambda.

Also adds a set of new functions for parsing fields with expected tags,
including a general purpose `tagged` for expecting a particular tag, and
ones for the supported ASN.1 types presently used by the `pkcs8` crate.