der: `Decodable` trait and core types
Refactors the decoder using a trait-based approach, centered around a
`Decodable` trait which operates over the initial extraction of a
`Decoder` type.

Adds structs for representing the ASN.1 core types we support, and then
impls the `Decodable` trait on those types.

Also uses the `Decodable` trait for decoding `Tag`, `Length`, and
`Header`, making it something of a universal abstraction.

Finally, impls the `Decodable` trait on types in the `pkcs8` crate.