der: add `Encodable` and `Message` traits (#159)
Replaces the previous function-based encoding implementation with a
trait-based approach, as companion work to the similar `Decodable`
trait introduced in #157.

The `Encodable` trait can be used for any of the parts of an ASN.1 DER
message, including the `Tag`, `Length`, and `Value`.

A higher-level `Message` trait is intended for encoding structs as ASN.1
`SEQUENCE`s. It provides a blanket impl of the lower-level `Encodable`
trait which only requires implementing a visitor for a message's fields.