der: extract `ByteSlice` type
Adds a newtype for byteslices that provides a central place to check the
`Length::max()` invariant inforced by this library.

Uses this newly introduced type to impl `Any`, `BitString`,
`OctetString`, and `Sequence`.

This eliminates the last of the `Result::expect` calls in this library
and moves it closer to being panic-free.