der: add `RawInteger` type
Adds a type which provides raw access to the underlying bytes of an
ASN.1 INTEGER which can be used for decoding/encoding "big integer"
types such as cryptographic keys.

Note that this implementation does not ensure that the encoding of such
integers is canonical. For now it's documented that the user of the
`RawInteger` type must perform such canonicalization.

This can potentially be addressed in a follow-up PR providing a generic
implementation of DER canonicalization rules.