der: handle leading byte of BIT STRINGs (#193)
The first octet of a BIT STRING encodes the number of unused bits.

Previous versions of this crate trated this octet as a verbatim part of
the bit string rather than decoding/encoding it.

This commit changes that behavior, presently ensuring the value is 0
(i.e. no unused bits) when decoding, and encoding a value of 0.