Rename `b64ct` => `base64ct`
Per #220, it seems like we need more general constant-time Base64
support:

- PHC "B64" for password hash
- PKCS#8 PEM encoding
- JWK (uses Base64url encoding)

The latter two are (potentially) private key use cases, so they're the
ones that would benefit the most from a fully constant-time
implementation.

The `b64ct` name comes from a specialization for the first use case.
However, now we're using `b64ct` for the latter two cases as well.

This commit renames the crate to `base64ct` with the goal of it covering
all of our use cases.