der: UTCTime and GeneralizedTime support
Adds support for two ASN.1 datetime types:

- `UTCTime` (as a `UtcTime` struct)
- `GeneralizedTime`

Both types are implemented as newtypes around a `Duration` representing
the amount of time since `UNIX_EPOCH` so as to support `no_std` targets
that do not have `SystemTime`.

Since `UTCTime` and `GeneralizedTime` are fundamentally "calendar"
types, this involves a small degree of calendar date support. The code
used was extracted from the `humantime`[1] crate, then subsequently made
stricter and minified so as to support `no_std`.

Encoding/decoding trait impls for `SystemTime` are also added, mapping
it to `GeneralizedTime`, which can support a much larger date range than
`UTCTime`.

Specific time handling follows the maximally strict "SHALL" directives
in RFC 5280, namely all timestamps are `Z`-normalized (or rejected), and
for `UTCTime` the 2-digit year interpretation of 1950-2050 is used.

[1]: https://github.com/tailhook/humantime