spki: add (back) an `AlgorithmParameters` type
Adds an enum representing `ANY`, `NULL`, and `Objectidentifier` cases
for the `parameters` field of `AlgorithmIdentifier`.

Ideally this enum would be unnecessary and we could just use the
`der::Any` type for this field, however to make that possible in all
relevant use cases we'd need a reference-based type for
`ObjectIdentifier` backed by a byte slice containing the DER
serialization of the OID. See:

https://github.com/RustCrypto/utils/issues/266

Until we have such a type, this enum suffices to make it possible to
decode and encode an owned OID value which stores the decoded arcs that
comprise it, rather than the DER serialization.