der: add ErrorKind::OidInvalid
Adds an error variant for propagating OIDs as discussed in #272.

It's a little awkward for a few reasons:

- There's already a unit `ErrorKind::Oid` variant. Since that's a little
  ambiguous, there's also a TODO to rename that variant to the more
  specific `ErrorKind::OidMalformed` in the next breaking release
- Since `const-oid` is an optional dependency, the presence of this
  variant is conditional upon that feature being enabled. Fortunately
  the whole enum is marked `#[non_exhaustive]` (which is what even
  allows adding new variants in a non-breaking way), so hopefully the
  conditional presence of this variant won't be too problematic for
  downstream users

The other problem is the `ObjectIdentifier` type is somewhat large. On a
64-bit machine, this increases the size of the overall `Error` type
(which wraps `ErrorKind` with additional context) from 32-bytes to
64-bytes.

That's not great, but probably worth it for the additional context being
able to propagate an OID as an error provides.