const-oid: fix root arc calculation
It was using bitwise OR instead of addition.

This commit also significantly expands the tests, including adding test
cases for a root arc of 2, assertions for the root arcs everywhere (as
they are represented differently internally), and additionally moves
tests into the `tests/` directory so we don't need conditional `std`
linkage for the tests.