der: add `Choice` trait (#295)
Closes #271

The `Choice` trait represents an ASN.1 `CHOICE` and allows types to
specify which ASN.1 tags they support.

This is implemented in terms of a static `can_decode` method. Using a
method for this means it's possible to impl `Choice` for `Any` (or e.g.
enums that contain `Any` as a member).

A blanket impl of `Choice` is provided for all types which impl
`Decodable + Encodable + Tagged`, i.e. types which can be viewed as a
`Choice` between only one alternative.

This trait is also an ideal for a custom derive macro for enums (and
indeed, it's necessary to support an `OPTIONAL CHOICE` which maps to
`Option<MyEnum>`.