const-oid: const-friendly string parser
Adds a new `ObjectIdentifier::parse` function which is declared
`const fn` and capable of parsing OIDs from dot-delimited strings
in const contexts.

This is much more readable/ideal than the current approach used by
`ObjectIdentifier::new`, which accepts a slice of `Arc` (i.e. `u32`)
type instead.

Ideally we can phase out the current `::new` in favor of one which
parses strings, however that'd be a breaking change, so for now the
implementation is side-by-side with the current `::new`.