der: make From<ObjectIdentifier> conversion for Any infallible (#319)
The `ObjectIdentifier` type has a small maximum length when serialized
as BER/DER, presently defined as `const_oid::MAX_LENGTH`.

This means that unless this were raised significantly such that it were
larger than `der::Length::max()`, infallible conversions should always
be possible. Such conversions are significantly more convenient,
especially in code that heavily leverages OIDs (e.g. the `pkcs5` crate)

This change opts for a potential panic in the event that the invariant
`const_oid::MAX_LEN <= Length::max()` were to be violated, but permits
an infallible conversion. Presently the invariant has the constants
`23 < 65535`. In the future it would probably make sense to *raise*
`Length::max`.

To ensure future changes/refactorings don't cause this invariant to be
violated, this commit includes a test that this invariant holds.