der: separate `sequence` and `message`
Closes #286

Adds a lower-level `SEQUENCE` API to `Encoder` which spawns a nested
encoder and passes it to a closure, similar to the corresponding
`Decoder` API.

Renames the previous "sequence"-related functionality which acts on a
slice of `Encodable` trait objects which represent the fields to
`message`.

Finally, removes `reserve` as part of the public API of `Encoder`, since
the new `SEQUENCE` API handles the previous use cases that required it.