der: make Length use u32 internally (#349)
The current restriction of `Length` being at most 65535 is actually just
an artifact of its inability to decode/encode lengths which are greater
than that limit, rather than an explicit design decision.

It would still be good to maintain a maximum document length, but
ideally it would be larger than 65535 (e.g. 128kB or 256kB).

This commit keeps the maximum length at its present value, but changes
the internal representation of the `Length` type to use `u32` to
futureproof it for a potential increase whenever the work is done to
support decoding/encoding larger lengths.