der: fix length overflow on add
In PR #349, the internal representation of a Length was changed from
`u16` to `u32` to make it easier to support documents larger than 64kB
in the future.

However, this added a `Length::max` invariant which wasn't getting
properly checked when doing addition (it was only checking that the
primitive `u32` integer addition operation didn't overflow).

This change adds a test which captured the overflow behavior, and also
adds a `TryFrom<u32> for Length` impl which ensures the length is
in-range, and changes the rest of the length-related code to always use
this impl to instantiate `Length` types.