der: make inner integer wrapped by Length fully private
Removes `pub(crate)` from the inner `u32` which the `Length` type wraps.

This ensures upholding the invariant that the inner length is less than
`Length::max` only needs to happen in the `length` module, and builds on
the work from #367 to ensure that `TryFrom<u32> for Length` is the one
place where this check occurs.