cpufeatures: support compiling on non-Linux/macOS aarch64 targets
ARM CPU feature detection is unfortunately tied to OS APIs, as the
registers containing information about CPU features are not accessible
to unprivileged userspace programs.

However, the previous setup which gated support on Linux/macOS only
means that OS-specific gating needs to be repeated in every downstream
consumer of `cpufeatures`, and also that if support for new environments
is added in `cpufeatures` that downstream consumers don't receive it
automatically.

This commit adds a fallback for non-Linux/macOS aarch64 environments,
similarly to how we handle SGX, which always returns `false` unless the
corresponding `target_feature`(s) are explicitly enabled in `RUSTFLAGS`.
This allows `cpufeatures` to provide the mechanism for
`target_feature`-based gating on these other platforms, even if it can't
provide runtime feature detection support.