der: support `Length` of up to 1 MiB
Closes #370.

Adds support for 4-byte DER-encoded lengths.

While this could potentially support lengths up to 16 MiB (2 << 24),
the library maintains a more conservative cap of 1 MiB as this should
be sufficient for the security-oriented cases this library is intended
to be used for.

This can be re-examined if there is a legitimate use case.