crypto-bigint
This commit adds an experimental prototype const generics-based big
integer library targeting cryptographic use cases.

It's an extraction of some of the code from the `elliptic-curve` crates,
modified to use const generics.

The immediate intended use case is being able to use it for elliptic
curve constants, such as a curve's `Order`:

https://docs.rs/elliptic-curve/0.9.11/elliptic_curve/trait.Order.html

As such, it's been written with liberal usage of `const fn`. This allows
safe usage in const scenarios which isn't possible with `generic-array`.

All that said, in general I think we could benefit from a library like
this for several applications, including things like `rsa` and `dsa`.