pkcs8: add TODO for proper extension handling; fix dups
The ellipsis `...` in the ASN.1 schema was previously interpreted as
allowing duplicate fields for `attributes` and `public_key`, however it
seems that it should correctly be interpreted as an extension marker
which allows extension fields to exist after `attributes` and
`public_key`.

It seems that, properly handled, extension fields should be ignored,
which we're not doing yet, so this is marked as a TODO for now.
The `der` crate should likely provide a better mechanism for skipping
over unknown extension fields.