der: add ContextSpecific field wrapper; fix BitString -> Any
Adds a `ContextSpecific` field wrapper type which decodes/encodes the
context-specific tag (sans the tag class prefix bits and constructed
bitflag) and inner value as an `Any`.

Additionally fixes a bug (discovered via the PKCS#8 impl) in the
`From<BitString<'a>> for Any<'a>` conversion which would result in the
leading `0` byte in the BIT STRING being absent. This involved a bit of
a special case hack in the way `Any` handles this conversion, where the
`Any` type automatically adds/removes it.

This hack could potentially be generalized into something more useful
(e.g. generalized support for a leading tag byte on any ASN.1 value
handled independently from the rest of the byte slice), but for now
it's handled as a special case.

Separately from the rest of that, this includes some overall code
quality improvements including making the internals of `Any` opaque
(important if it's handling special cases), and fixes for various
infallible conversions which previously performed needless fallible
conversions followed by `.expect(...)`.