cpufeatures: add iOS support (static ARM64 capabilities only)
All Apple ARM64 hardware has the same baseline set of statically known
capabilities which can be assumed on all iOS (and macOS) platforms.

This commit adds support for those statically known capabilities on iOS.

Unfortunately it does not appear to be possible to access the
`sysctl(3)` namespace on iOS in order to determine the availability of
other CPU features which aren't part of this baseline set the same way
we can on macOS, so a static capability set is the best we can do.

See this issue for more information:

https://github.com/RustCrypto/utils/issues/378