crypto-bigint: optional `generic-array` interop
There is one interop strategy that is possible between `generic-array`
and const generics today, and it's similar to the one already used by
`typenum` (and, as it were, `crypto-bigint` itself): avoid any sort of
const-generic parameterization when dealing with traits, and provide an
impl of a trait for each desired size, with a macro to write them.

This commit adds a (currently decode-only) `ArrayEncoding` trait with a
`ByteSize` parameter which must be an `ArrayLength<u8>`.

It uses the macros for defining type aliases for `UInt` to add an impl
of this new trait.

As it were, `typenum` doesn't support all of the sizes that were
previously defined, however it seems like the sizes that `typenum` does
support are actually the sensible ones for a cryptographic bignum
library, so this commit also matches the `U*` sizes with the ones
provided by typenum.