crypto-bigint: `mul_wide`/`wrapping_mul`/`checked_mul` for UInt
Implements schoolbook multiplication of two `UInt` values, with a core
"wide" multiplication operation which is twice the width of the input
value, currently implemented as a 2-tuple and aliased to `Wide`.

Also provides `wrapping_mul` and `checked_mul` methods composed from the
wide multiplication method which either discard overflow or detect and
fail on overflow.