crypto-bigint: add `Wrapping` wrapper type
Adds a type analogous to `core::num::Wrapping` which always performs
wrapping arithmetic on the inner value.

Initial support is provided for `UInt`, but like the `core`/`std` type
it's designed to be generic and provides some generic implementations as
well, in case we ever add a signed integer type to this library.