crypto-bigint: add `nlimbs!` macro
Adds a macro which can compute the number of limbs required to store an
n-bit number at compile time.

This is intended as a stop gap until a forthcoming Rust feature like
`const_evaluatable_checked` allows it to be replaced with something such
as a `const fn`.

For now it's needed to compute various const generic parameters however,
so a macro is the only option available.