cpufeatures: map `aarch64` HWCAPs to target features; add `crypto`
Closes #395

The HWCAPs provided by Linux are finer-grained than LLVM target
features. This crate is trying to detect the latter, so we need to group
together these finer grained features detected via HWCAPs.

This commit attempts to group together finer grained HWCAPs into coarser
grained target features, ensuring all of the relevant HWCAPs for a given
target feature are available.

Additionally, this adds support for the `crypto` target feature, which
implies `aes` (and with it PMULL), along with `sha2` and `neon`
(although the latter is always present on `aarch64`)