crypto-bigint: add `From` conversions between `UInt` and limb arrays
These will make it much easier to incrementally retrofit `UInt` into the
`elliptic-curve` crate and the respective curve implementations.

From an encapsulation perspective, it's a bit problematic because it's
exposing the internal representation used by `UInt`, so it additionally
adds TODOs to eventually remove these as they're a bit of a hack.