der: add `TagNumber` type, make `Tag` structured (#464)
Adds a `TagNumber` type for representing the lower 5 bits of a Tag,
where bits 8/7 identify the class and bit 6 identifies primitive vs
constructed forms.

Changes `Tag` from being `repr(u8)` to a structured representation with
`Application`, `ContextSpecific`, and `Private` variants, each of which
includes a `TagNumber`. This enables support for all of these classes
for tag numbers up to 30, which is the maximum value which can be
represented with a single leading identifier octet.

This change enables the `ContextSpecific` struct to have public members,
as the tag number invariant can now be maintained by the `TagNumber`
type instead.