der: const generic unsigned integer decoder with u32/u64/u128 support
This is another take on #325, but using const generics as a way to
provide a common implementation for Rust's unsigned integer primitives.

This commit replaces the existing DER decoder/encoder implementations
for integer primitives with a core const generic implementation, and
uses macros to write implementations of `TryFrom<Any>`, `Encodable`, and
the `Tagged` traits for each of the integer primitives.

This commit extends support to `u32`, `u64`, and `u128` using:

    impl_uint_encoding!(u8, u16, u32, u64, u128);