crypto-bigint: add UInt::into_limbs
The intended use case is initializing the existing `MODULUS` constants
for the various `Scalar` impls in the elliptic curve crates.

Longer term we should be able to migrate to an entirely `UInt`-based
implementation of scalar (ideally adding reusable `Field` types as
well), but for now this is being added to simplify the initial retrofit.