crypto-bigint: `Limb` newtype (#499)
Adds a `Limb` newtype for the integer type which represents a limb for a
given CPU word size.

This allows us to provide much more consistent properties around what
represents a "limb", as well as define functions directly on the limb
type, which makes it much easier to define e.g. carry/borrow chains.