pem-rfc7468: initial crate
Adds an implementation of RFC7468-flavored PEM encoding.

Targets the "Strict" encoding exclusively (for now), attempting to
adhere to the conventions set out in RFC7468.

The implementation is carefully written to avoid branching or table
lookups on any data which might potentially contain a secret, making it
constant time with respect to secret data. It also avoids copies or any
usage of temporary buffers. Heap usage is optional, and the
implementation can therefore function in "heapless" no_std environments.

Some parts of the implementation are currently overly strict even
according to the strictest interpretation of RFC7468 and should be
relaxed to ensure conformance.

The intended goal of splitting this out into its own crate is to use it
for PEM encoding in the `pkcs1`, `pkcs8`, and `spki` crates.