pem-rfc7468: increase LF precedence in EOL stripping functions
CRLF handling can potentially touch secret data.

Unfortunately in the trailing CRLF case this is somewhat unavoidable,
as looking at the previous byte before a LF is needed to disambiguate
CRLF from LF.

However, for a leading EOL, if LF is examined first we can skip the CRLF
check, and avoid branching on potentially secret data.

This should reduce the total number of secret bytes potentially branched
on in the happy path to just 1-byte: handling the EOL that separates
the encapsulated text from the post-encapsulation boundary.