der: fix encoding for 20xx UTCTime dates
Closes #568

UTCTime uses a 2-digit year, ranging from 1950 to 2049.

Encoding for years after 1999 was being mishandled, because they rolled
over to 100, resulting in a three digit year when only 2 digits are
supported by the encoder.

This commit ensures dates with years 2000 to 2049 are encoded as 0..=49,
and adds tests similar to the ones in `DateTime` which ensure samples
within the entire supported range round trip.