base64ct: make `Encoding::decode` reject invalid padding (#577)
When `Variant::PADDED` is `true`, ensure that the decoded data
round-trips back to the padded input. If the padding is malformed,
returns `Error::InvalidEncoding`.

Note that this commit only impls this check for the non-in-place
`decode` method, not for `decode_in_place` yet. Instead, it documents
that `decode_in_place` does not yet perform this check, and leaves it
to be implemented in a followup. However, it does document that the
`decode_in_place` method does not (yet) perform this check.