[WIP] crypto-bigint: `add_mod` discrepancy
For an as yet undetermined reason I can't implement scalar arithmetic
for the `k256`/`p256` crates using the current `add_mod` implementation
(or rather, the test vectors fail).

It appears to have something to do with overflow handling, and how the
current implementation unconditionally subtracts the modulus regardless
of whether or not the add operation overflowed.

However, while the implementation in this commit allows the
`k256`/`p256` tests to pass, it fails all of the current proptests.