crypto-bigint: fix `add_mod` overflow handling (#619)
This PR is effectively an extraction of modular addition as implemented in
the `p256` crate.

It conditionally subtracts the modulus on overflow.

It includes a test case extracted from `p256` which captures a case
where the previous `add_mod` implementation was miscomputing results.

It seems the previous test cases were somewhat problematic, with test
failures for cases which shouldn't have been tested. They have been
removed for now.