[WIP] flex-array: hybrid typenum/const-generic arrays
This crate is a `generic-array`-inspired array type which uses some of
the ideas from `crypto-bigint` to provide type-level conversions between
const generics and typenum-based arrays.

I originally looked at trying to add this sort of thing as an
afterthought to `generic-array`, but felt that it would really be
helpful if it were the core abstraction everything else is built on.
Also `generic-array` is full of huge swaths of complicated `unsafe` code
which is completely unnecessary in the setup used in this crate, which
merely uses the type system to map to Rust's underlying array types.

It provides a `FlexArray` type similar to `GenericArray` aloing with an
`Array<T, const N: usize>` trait which can be used to map const generic
parameters to underlying array types.

Macros are used to write the `Array`/`ArrayLength` trait impls along
with other traits which can't be implemented generically because of
unconstrained generic parameters: `Index`/`IndexMut`, `IntoIterator`,
and `TryFrom<&[T]>`. These are impl'd for the following sizes.

- 0-64
- 96
- 128
- 192
- 256
- 384
- 448
- 512
- 768
- 896
- 1024
- 2048
- 8192

This is a smaller range of sizes than what `generic-array` supports due
to `typenum` type alias definitions. It is necessary to make the
const generic linkage work.

Compile times are similar to `generic-array` with the current range of
supported types. In `debug` mode I get 4s for `flex-array` and 2s for
`generic-array`, and in `release` mode I get the opposite with 2s for
`flex-array` and 4s for `generic-array`.

More sizes could be added at the cost of more compile time. Likewise
there are some self-referential relationships between the traits that
could be removed to improve compile times, but seem handy.

The overall API is *much* closer to `core::array` than `GenericArray`
is. For example, `GenericArray` has a panicking `From` providing
`TryFrom`-like behavior, and confusingly receives the blanket impl of
`TryFrom`, so people can mistakenly use `TryFrom` expecting a fallible
API only to get the blanket impl of `From` and a panic in the event of
an error.

Everywhere possible, it uses const generic implementations of traits.
This includes:

- `AsRef<[T; N]>`
- `AsMut<[T; N]>`
- `Borrow<[T; N]>`
- `BorrowMut<T; N]>`
- `From<[T; N]>`