Use as_mut_ptr to get a pointer for mutation (#728)
The documentation for `slice::as_mut_ptr` says:
The caller must also ensure that the memory the pointer (non-transitively)
points to is never written to (except inside an UnsafeCell) using this pointer
or any pointer derived from it. If you need to mutate the contents of the
slice, use as_mut_ptr.