Initial `cmov` crate
Implements safe wrappers for the `cmovz` and `cmovnz` instructions which
work on `x86` and `x86_64`, implemented using inline assembly.

Using inline assembly allows us to provide a sort of black box which
LLVM will not interfere with, which is otherwise problematic when using
anything besides ASM.

It's otherwise not possible to correctly emit CMOV instructions on x86
platforms with LLVM because the `x86-cmov-conversion` pass which will
potentially rewrite them as branches.

For more details, see:

https://dsprenkels.com/cmov-conversion.html