hybrid-array: hybrid typenum/const-generic arrays (#707)
This crate provides a `generic-array`-inspired array type which uses some
of the ideas from `crypto-bigint` to provide type-level conversions between
const generics and typenum-based arrays.

I originally looked at trying to add this sort of thing as an
afterthought to `generic-array`, but felt that it would really be
helpful if it were the core abstraction everything else is built on.
Also `generic-array` is full of huge swaths of complicated `unsafe` code
which is completely unnecessary in the setup used in this crate, which
merely uses the type system to map to Rust's underlying array types.

It provides an `Array` type similar to `GenericArray` along with an
`Array<T, const N: usize>` trait which can be used to map const generic
parameters to underlying array types.

Macros are used to write the `Array`/`ArraySize` trait impls for the
following array sizes:

- 0-64
- 96
- 128
- 192
- 256
- 384
- 448
- 512
- 768
- 896
- 1024
- 2048
- 8192

This is a smaller range of sizes than what `generic-array` supports due
to `typenum` type alias definitions. It is necessary to make the
const generic linkage work.

The overall API is *much* closer to `core::array` than `GenericArray`
is. For example, `GenericArray` has a panicking `From` providing
`TryFrom`-like behavior, and confusingly receives the blanket impl of
`TryFrom`, so people can mistakenly use `TryFrom` expecting a fallible
API only to get the blanket impl of `From` and a panic in the event of
an error.

Everywhere possible, it uses const generic implementations of traits
rather than macro-based implementations in order to improve compile
times and keep the macro-based parts to a minimum.

This includes:

- `AsRef<[T; N]>`
- `AsMut<[T; N]>`
- `Borrow<[T; N]>`
- `BorrowMut<T; N]>`
- `From<[T; N]>`
- `Index<I>`
- `IndexMut<I>`
- `TryFrom<&[T]>`