cpubits: initial crate
As discussed in #824, adds a crate with the intent of it providing
heuristics for selecting the optimal word size to use for a particular
target CPU, which may differ from its address size.

It's implemented as `macro_rules` to avoid a build script.

Currently there's no ability to override it, though we could consider
adding something like `cfg(cpubits = "64")`.