hybrid-array: add sealed::ArrayExt trait
Adds an extension trait with a blanket impl for all core arrays, and
adds it to the bounds for `ArraySize::ArrayType`.

It currently provides a `from_fn` constructor which is now leveraged by
a now-generic `Default` impl which bounds on `ArraySize`.

This reduces the bounds needed in generic code to just `ArraySize`, even
to leverage the `Default` impl, and should also permit some additional
array constructors.