hybrid-array: add `FromIterator` impl
For feature parity with `GenericArray`.

It would be nice to have a fallible version of this which returns an
error instead of panicking on length mismatch, but we'd first need to
add `FromFn::try_from_fn` or thereabouts.

This is enough to cover the immediate use cases in the meantime.