fiat-constify: add support for leveraging `const_mut_refs` (#1177)
The generated code is *almost* `const fn` friendly with `const_mut_refs`
but needs `const fn` accessors for the newtypes it defines:

https://github.com/mit-plv/fiat-crypto/issues/1955

This updates `fiat-constify` to solve only adding such accessors and
annotating the functions as `const fn`, greatly reducing its scope and
making the resulting postprocessed code much closer to the original
generated code.

Additionally it removes the extra methods added to the code, namely
`as_inner` and `into_inner`, in order to ensure that the output is as
close to the original generated code as possible.