ctutils: constant-time selection and equality testing
Inspired by the `subtle` crate, this is a next-generation constant time
utility crate built on the `cmov` crate's constant-time
selection/predication and equality comparisons, which are exposed as
`CtSelect` and `CtEq` traits (equivalent to `ConditionallySelectable`
and `ConstantTimeEq` in `subtle`).

Additionally, it uses `black_box` as a "best effort" optimization
barrier whenever accessing its constant time conditional type (called
`Choice` as in `subtle`), providing an additional line of defense
against possible future compiler optimizations.

It also has a `CtOption` type like `subtle` which provides a
constant-time equivalent to `Option` but with combinators that are
evaluated eagerly rather than lazily so they behave the same regardless
of the effective absence or presence of the underlying value.

It proactively makes functions `const fn` wherever possible, making it
possible to construct and use `Choice` and `CtOption` in these contexts.

The `Copy` bounds have been removed, making it possible for everything
to be used on heap-allocated types, such as `BoxedUint` in
`crypto-bigint`.

The above two issues taken together are the main reasons why
`crypto-bigint` currently embeds its own mini-`subtle` alike, which it
would be nice to use this library to replace.