ctutils: downgrade MSRV to 1.85 (#1259)
This is implemented as its own PR so we can easily revert it when we are
ready to bump MSRV to 1.87.

The main thing we lose here is `const fn` support for `Choice::to_u8`
and `to_bool`, since these use `black_box`.

In addition the macros to write the signed impls of `CtEq` and
`CtSelect` become a little uglier, but that's not the end of the world.