ctutils: import `Choice` methods from `crypto-bigint` (#1266)
This imports all of the `const fn` constructor and predication/selection
methods `crypto-bigint` needs to replace `crypto_bigint::ConstChoice`
with `ctutils::Choice`, along with extant tests.

To make everything work, at least for now, a `word` feature has been
added to `ctutils` in order to have methods on `Choice` which operate on
`Word` and `WideWord` that use the same definition as `crypto-bigint`.