ctutils: impl `From<u8>` for `Choice`
This is the main way `subtle` constructs `Choice` so it's important for
backwards compatibility.

It seems like if we had `from_u8_lsb`, we could avoid the panic
condition by always masking the input `u8`. We don't yet, but this
commit also leaves a comment to consider deprecating both this
`From<u8>` impl and `Choice::new` in the future.